Implement codebase labeling
And a few example rules:
- package.json present => deps:node label
- jest dependency declared => test:jest label
- go.mod present => deps:go label

The order of rules matters, as some rules
could depend on others. E.g. only check for
jest if deps:node has matched.

The of the MatchData structure is that labels
can carry some metadata specific to them.

In particular "deps:..." labels are generated
when matching package manager and/or task
running tooling and we want to include the
dependencies and tasks defined for those tools.