Implement structs representing a config
And methods that dump them to yaml.Node.

Not all options are implemented, but should be
sufficient for most inferred configs.

I considered using textual templates instead
which might have been easier to implement,
give us a little bit more flexibility in some
cases (whitespace, comment position), and
could be easier to edit for people not familiar
with Go, but I think the advantage of
composability of the yaml.Node structures
trumps those considerations.