Implement config generation framework
Added some simple jobs for Node and Go.
They will need further refinement.

If the Jobs generated depend on Orbs, they are
included in the config.

At the time being all jobs are added to a
single workflow with no dependencies between
them. But the idea is that we can later
be a bit smarter about this and make e.g. the
artifact build jobs depend on the test jobs.
I have defined the job type for this purpose.

There is also a stub fallback workflow.
Pending decision on what exactly to include there.