Only walk the file tree once
Of course, that means we always have to walk the whole tree.
But that's going to be the case for almost all codebases, anyways.

Also stop after 250k files.

There's probably room for further optimizing this, by e.g. calling
filepath.Base on each file only once, or perhaps stopping at a certain
dir depth, but this at least puts an upper bound on execution time.